+++ aztec/src/main/kotlin/org/wordpress/aztec/formatting/BlockFormatter.kt
@@ -925,7 +925,7 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
                         }
                     } else {
                         (selStart in spanStart..spanEnd || selEnd in spanStart..spanEnd) ||
-                                (spanStart in selStart..selEnd || spanEnd in spanStart..spanEnd)
+                                (spanStart in selStart..selEnd || spanEnd in selStart..selEnd)
                     }
                 }
     }